Moved more classes to the api package
