+++ src/main/java/media/xen/tradingcards/api/card/CardInfo.java
@@ -1,0 +1,10 @@
+package media.xen.tradingcards.api.card;
+
+import lombok.Data;
+
+@Data
+public class CardInfo {
+    private final String name;
+    private final String colour;
+    private final String display;
+}
